`AnyNumeric` is no longer backed by Postgres-allocated memory
Until now, `AnyNumeric` (and by extension, `Numeric`) has been backed by a Postgres-allocated pointer to `pg_sys::NumericData`.  This pointer would have been allocated by the `CurrentMemoryContext` and could become freed if that memory context were to be freed prior to the Rust `AnyNumeric` wrapper going out of scope.

`pg_sys::NumericData` is actually a "varlena" in disguise, and this allows us to copy the backing bytes into Rust-owned memory, which is then properly managed for the life of the `AnyNumeric` wrapper.

Conversion from an `AnyNumeric` back to the generic `Datum` is simply a matter of copying that data into Postgres-allocated memory.