Ignores UI tests for MUSL environments (#1235)
The new UI tests introduced in `develop` seem to be incompatible with
musl environments such as Alpine. The current working theory is that
there could be a bug in `TryBuild` that ignores/replaces certain rustc
flags. In particular, it seems to be disregarding the
`"-Ctarget-feature=-crt-static"` flags that are necessary for pgrx to
build and be used in Alpine.

Also includes a couple minor updates, such as bumping the
postgresql-alpine Docker version.