Addresses cargo-pgrx error reporting (#1238)
There are some situations where `cargo pgrx` reports an error but
important details are missing, as it relates to the spawning of child
process. One example of this is when `cargo pgrx` downloads and compiles
PostgreSQL from source, as demonstrated by
https://github.com/pgcentralfoundation/pgrx/issues/1222

The changes here seek to make any error output due to a spawned child
process be reported to the end user instead of being silently swallowed
and disregarded.