Use BTreeSet for consistent ordering
This removes the is_for_release being permeated through bindgen,
into parts which don't need to know about that step, and removes
a source of inconsistency in our build artifacts.

It also avoids having to do an additional reallocate-and-sort step,
by simply using a type that was always going to be sorted anyways.