Pedantically use `pg_sys::Datum` (nfc) (#1372)
Use `pg_sys::Datum` consistently throughout pgrx. This brings pgrx
better in conformance with its own style, and also is an important
refactor for reasons that will become obvious in a later PR. No
functional changes.

In a few places I also touched up some other imports, but that isn't
done as consistently, so I have no idea if this misses a spot on that.