Allow safe Oid conversions
Simply cast `u32 as i32` and pass it through SPI to do this by other means.
While it is an incredibly bad idea to do so, unfortunately that means we
cannot rely on the conversion being `unsafe`, as there is no way we would
make SPI unsafe. That leaves no persuasive argument for not providing From.