Make `#[derive(PostgresType)]` impl its own FromDatum (#1381)
Mainly, this removes a source of persistent and confounding type errors
because of the generic blanket impl of FromDatum for all T that fulfill
so-and-so bounds. These may mislead one, if one is writing code generic
over FromDatum, to imagine that one needs a Serialize and Deserialize
impl or bound for a given case, even when those are _not_ required. By
moving these requirements onto the type that derives, this moves any
confusion to the specific cases it actually applies to.

This has a regrettable effect that now PostgresType _requires_ a
Serialize and Deserialize impl in order to work, _unless_ one uses the
hacky `#[bikeshed_postgres_type_manually_impl_from_into_datum]`
attribute, which I intend to rename or otherwise fix up before pgrx
reaches its 0.12.0 release.