Sniff for libclang-version-correct headers (#1325)
Bindgen's strategy of autodetecting headers attempts to use a hacky
route used by clang-sys, however, it doesn't take into account that the
strategy of clang-sys uses `$CLANG_PATH` instead of `$LIBCLANG_PATH`,
and then falls back to searching the entire directory structure. This
means it's fairly easy for it to pull the headers for a different
version, trying to compile them with the wrong clang!

This replaces their series of unfortunate hacks with our own, using
the collected information from pg_config, clang-sys, and rust-bindgen
to try to locate the correct compiler headers. If it can't find anything,
it falls back to doing what currently happens.