Enable ignoring some commits in blame
In general it is bad form to remove commits from the commit history,
however some commits shed more heat than light when examining blame.
Add a file to assist with ignoring them.

This name in particular is conventionally used by code forges, so
it will be used for their webview as well.