Prepare pgrx v0.11.2 (#1425)
pgrx v0.11.2 is a minor release which

- makes available the storage-related API, thanks to @silver-ymz in
https://github.com/pgcentralfoundation/pgrx/pull/1409
- deprecates the `Oid::from_u32_unchecked` API... because it is actually
possible to do via casting in a *query*, which makes it effectively
impossible for us to not wind up providing it via some other safe API,
since the source is also an easily-Copied type. Thus you can now simply
use `From::from` for it. Thanks to @thomcc for this discovery,
implemented by @workingjubilee in
https://github.com/pgcentralfoundation/pgrx/pull/1374

As usual, `cargo install cargo-pgrx --version 0.11.2 --locked` and be on
your merry way!

Thanks everyone!