Manually staticize lifetimes in pgrx tests
These all are actually lifetimes bound by their invocation site,
which means that they return any lifetime which is named, instead
of a lifetime that is bound by the input types in a logical way.
This means that these lifetimes might as well be 'static, for all
it would (not) matter.