Use heapless 0.8 (#1495)
Note this is an API break for users of heapless because we expose this
in our own public API and they need to match the types and such. We
should probably constrain this problem more by reexporting the relevant
crates, so pgrx's users can simply use the reexported types and traits.