Allow pgrx to be slightly messy (#1527)
As noted in #1517 there are many distractions in this repo due to some
people using cargo clippy or other lints. For some of them, I do not
intend to fix them, or intentionally am deferring fixing the lint
(because the lint is triggering due to a genuine problem, and the
genuine problem should be fixed, instead of the smaller change for
"fixing the lint").