Remove unused `#[pgrx(..)]` support
The support for this attribute in other cases, with other styles of arg,
was never fully implemented. It seems `#[pg_aggregate]` handles the
only case of this happening in the codebase on its own? The name = val
style is the only case actually handled by this code.

Remove unused code as it complicates debugging and updating the rest,
especially with the lack of testing for its cases. Also touch everything
with Debug implementations, so that will be easier in the future.