Reuse the same attr parsing scheme
Considering only the inner args of the attribute is necessary
in order to get the correct names so that we can emit correct SQL.
Ideally we would have much more resilient parsing that could
extract data out of a broader range of token streams passed in.
However, that is not the case currently, so don't.