add bindings of subscript function (#1562)
Add bindings of supporting subscript access for user-defined types, eg.
`udf_type_value[1]`. It's supported since PostgreSQL 14.

Ref: https://www.postgresql.org/docs/current/sql-createtype.html

> The optional `subscript_function` allows the data type to be
subscripted in SQL commands. Specifying this function does not cause the
type to be considered a “true” array type; for example, it will not be a
candidate for the result type of `ARRAY[]` constructs. But if
subscripting a value of the type is a natural notation for extracting
data from it, then a `subscript_function` can be written to define what
that means. The subscript function must be declared to take a single
argument of type `internal`, and return an `internal` result, which is a
pointer to a struct of methods (functions) that implement subscripting.
The detailed API for subscript functions appears in
src/include/nodes/subscripting.h. It may also be useful to read the
array implementation in src/backend/utils/adt/arraysubs.c, or the
simpler code in contrib/hstore/hstore_subs.c. Additional information
appears in [Array Types] below.

[Array Types]:
https://www.postgresql.org/docs/current/sql-createtype.html#SQL-CREATETYPE-ARRAY

---------

Signed-off-by: usamoi <usamoi@outlook.com>