Rename `ErrorReportable::report()` to `unwrap_or_report()` (#1568)
As noted in #1561, `pgrx::pg_sys::panics::ErrorReportable::report()` is
named identically to `std::process::Termination::report()`, and they are
both traits that are implemented on `Result<T, E>` so if you import both
traits you'll have a name collision. Aside from the name collision, it's
very easy to see `.report()` called on a `Result` type and make the
incorrect assumption this is the `Termination` trait when one is
unfamiliar with PGRX's codebase.

To avoid ambiguity, this set of changes renames that method to
`unwrap_or_report()`, and is intended to be included in the set of
breaking changes for 0.12.