pfree toasted Arrays on drop (#1571)
For tables with many rows, or that include enough large arrays, being
unable to pfree the entire array can cause needless memory pressure. As
pgrx used to not bind objects to the lifetime of backing arrays, the
assigned lifetime of any 'obj could exceed `&'arr Array<'_, T>`. This
rendered any Drop implementation for Array or any of its parts unsound.
The references, expecting to be valid for the underlying 'mcx or even
longer, would yield clobbered memory when actually examined.

This has finally become untrue with the introduction of UnboxDatum. It
has an `unsafe trait` obligation that the implementation not allow
lifetimes to exceed the Datum's lifetime except if the data is always
inherently 'static or Copy. This allows deallocating detoasted items, as
Arrays now only yield an object of an appropriate lifetime. It is
possible for us to get this wrong, but only by implementing UnboxDatum
in an incorrect, unsound way, so the fix would be to that code instead.