Remove the `pg_sql_graph_magic!()` macro.
The `pg_sql_graph_magic!()` did some shenanigans related to loading the extension ".control" file at compile-time and embedded it in the compiled artifact.  This was convoluted and actually unnecessary as the location where that information was used (in `schema.rs`) can easily
access the ".control" file at runtime.

This PR removes that macro and its usages.

The `fn __pgrx__marker()` function is still required, however.  It's necessary to ensure that the `rustc` build we do for the `pgrx_embed` binary for schema generation properly links so that the other (dynamically referenced) `__pgrx_internals_XXX` symbols can be found.  The function, however, is now a noop.