Do not take &mut to uninit data
More of the same, but in addition, we avoid taking &mut to what could be uninit,
and simply use palloc0 instead. This could hypothetically lead to perf regressions?
In actuality, however, it probably is faster, if anything, and to be sure, we now
instead elide the bounds check for every index of the args by directly iterating.